pyproject.toml
[project]
name = "mcp-logs-reporter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jinja2>=3.1.6",
    "mcp[cli]>=1.17.0",
    "psycopg[binary]>=3.2.10",
    "pydantic>=2.12.2",
    "python-dotenv>=1.1.1",
]
